Make package modular
