+++ pip_package/setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = tokenization-layer
-version = 0.0.1
+version = 0.0.2
 author = Martin Molnar
 author_email = martin.molnar07@gmail.com
 description = An NLP tokenization algorithm that is a trainable layer for neural networks.
